Last attempts to get standalone version fully working.
