+++ src/app/app.component.spec.ts
@@ -6,9 +6,9 @@ describe('AppComponent', () => {
   beforeEach(async () => {
     await TestBed.configureTestingModule({
       imports: [
-        RouterTestingModule
+        RouterTestingModule,
+        AppComponent
       ],
-      declarations: [AppComponent],
     }).compileComponents();
   });
